Implementation of TICKET-102.
This implementation enables the validation of customer country. A
country name must now be provided. If the country is outside the Baltic
scope, the loan will not be issued.

Additionally, age validation has been added. The loan will not be issued
if the customer is underage or older than the expected lifetime of each
respective country minus our maximum loan period.

New class have been created to support these validations.